Avoid saving/loading patches off by one
Prior to this change, plugging in the number 18 would load the patch Plinky says is at 19, and so on. Users see a 1-indexed set of patches from 1 to 32, and the software sees a 0-indexed set from 0 to 31.

I'm index shifting at the load & save points, so everywhere else in the app will refer to the patch number by the user-facing value. 

I've also put some upper and lower bounds on patch-numbers, as chrome doesn't seem to enforce the min & max properties on the number input. trying to load patch 0 will load patch 1, trying to load patch 1000000 will load patch 32, etc.